pyproject.toml
[project]
name = "rekuest-server"
version = "0.1.0"
description = ""
authors = [{ name = "jhnnsrs", email = "jhnnsrs@gmail.com" }]
requires-python = ">=3.12,<4"
readme = "README.md"
dependencies = [
    "django>=5.2",
    "kante==0.9.0",
    "omegaconf>=2.3.0,<3",
    "psycopg[binary]>=3",
    "daphne>=4.0.0,<5",
    "channels-redis>=4.1.0,<5",
    "django-choices-field>=2.2.2,<3",
    "authentikate==0.9.0",
    "namegenerator>=1.0.6,<2",
    "ujson>=5.8.0,<6",
    "aioredis>=2.0.1,<3",
    "jsonpatch~=1.33",
    "django-probes>=1.7.0,<2",
    "rich>=14.0.0",
]

[dependency-groups]
dev = [
    "django-stubs>=5.2.0",
    "mypy>=1.15.0",
]

[tool.django-stubs]
django_settings_module = "rekuest.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "rekuest.settings"


[tool.ruff]
# Enable the pycodestyle (`E`) and Pyflakes (`F`) rules by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
extend-select = ["ANN", "D1"]
extend-ignore = [ "ANN002", "ANN003", "D100", "ANN401"]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "tests",
    "examples",
    "**/migrations",
]

# Same as Black.
line-length = 300


# Assume Python 3.12.
target-version = "py312"


[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
upload_to_pypi = true
branch = "main"
build_command = "uv build"



[tool.mypy]
exclude = ["venv/", "tests/", "examples/"]
plugins = ["mypy_django_plugin.main","pydantic.mypy"]
ignore_missing_imports = true
strict = true
